Treat the --deps arg as a boolean, not a string
The --deps argument was being treated as a string, causing
the default 'FALSE' to thow an error.  It is now treated
as a boolean, like the --error toggle.

The R function install.packages allows deps to be either
a logical/boolean value or a string with any combination of
the elements SUGGESTS, IMPORTS, DEPENDS, etc, so this
change limits the functionality.  Trying to handle both
cases with command line arguments would probably make the
code too messy, so I just stuck with the boolean-only strategy.