Add additional options to installBioc.r
Adds additional options to the installBioc.r example script
based on the approach used in install2.r.

The main goal was to be able to use the installBioc.r script
as the main method to install bioconductor packages in Dockerfiles,
and to make it fail the build process whenever an error
occurs during package installation.

The script now throws errors when a package fails to install
due to missing dependencies, misspelled names, etc. Already
installed packages can be skipped. Suggested dependencies
can be toggled. Parallel install can be enabled.
Download method can be set.

A few options were added for BiocManager::install()'s
unique arguments. A toggle to update other packages that are
outdated. Additional repositories can be prepended to the
standard BioC repository (only works for bioc sub-repositories
like software/experiment). Already installed packages can be
forced to redownload and install.

In addition, the library path is now set in the same manner as
is done in install2.r (by reading the first entry of .libPaths()
and passing it to the install function, rather than setting it
 to `/usr/local/lib/R/site-library` via `.libPaths()`).